+++ src/Model/OpeningHours.php
@@ -6,6 +6,7 @@ use SilverStripe\Forms\CheckboxSetField;
 use SilverStripe\Forms\DropdownField;
 use SilverStripe\ORM\DataObject;
 use SilverStripe\Versioned\Versioned;
+use SilverStripe\Forms\TimeField;
 
 class OpeningHours extends DataObject
 {
@@ -20,8 +21,8 @@ class OpeningHours extends DataObject
 
     private static $db = [
         'Days'      =>  'Varchar(255)',
-        'TimeOpen'  =>  "Enum(',00:00,01:00,02:00,03:00,04:00,05:00,06:00,07:00,08:00,09:00,10:00,11:00,12:00,13:00,14:00,15:00,16:00,17:00,18:00,19:00,20:00,21:00,22:00,23:00,24:00', '')",
-        'TimeClose' =>  "Enum(',00:00,01:00,02:00,03:00,04:00,05:00,06:00,07:00,08:00,09:00,10:00,11:00,12:00,13:00,14:00,15:00,16:00,17:00,18:00,19:00,20:00,21:00,22:00,23:00,24:00', '')",
+        'TimeOpen'  =>  'Time',
+        'TimeClose' =>  'Time',
     ];
 
     private static $has_one = [
@@ -76,15 +77,13 @@ class OpeningHours extends DataObject
                         'So' => 'Sunday'
                     ]
                 ),
-                DropdownField::create(
+                TimeField::create(
                     'TimeOpen',
-                    'Opening Time',
-                    singleton(self::class)->dbObject('TimeOpen')->enumValues()
+                    'Opening Time'
                 ),
-                DropdownField::create(
+                TimeField::create(
                     'TimeClose',
-                    'Closing Time',
-                    singleton(self::class)->dbObject('TimeClose')->enumValues()
+                    'Closing Time'
                 )
             ]
         );